requirements.txt
PyMuPDF>=1.24.0
pdf2image>=1.17.0
pytesseract>=0.3.10
Pillow>=10.0.0

fastapi>=0.112
uvicorn>=0.30
python-multipart>=0.0.9
